+++ package.json
@@ -16,13 +16,13 @@
   "license": "MIT",
   "author": "Alfi Maulana <alfi.maulana.f@gmail.com>",
   "exports": {
-    "./env": {
-      "default": "./dist/env.js",
-      "types": "./dist/env.d.ts"
-    },
     "./exec": {
       "default": "./dist/exec.js",
       "types": "./dist/exec.d.ts"
+    },
+    "./io": {
+      "default": "./dist/io.js",
+      "types": "./dist/io.d.ts"
     },
     "./log": {
       "default": "./dist/log.js",
