+++ typedoc.json
@@ -1,4 +1,4 @@
 {
-  "entryPoints": ["src/env.ts", "src/exec.ts", "src/log.ts", "src/vars.ts"],
+  "entryPoints": ["src/exec.ts", "src/io.ts", "src/log.ts", "src/vars.ts"],
   "treatWarningsAsErrors": true
 }
